PySimpleGUI_PyWebIO_Streamlit/requirements.txt
altair==4.1.0
altgraph==0.17
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bar-chart-race==0.1.0
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
japanize-matplotlib==1.1.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.2
numpy==1.21.0
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
plotly==5.1.0
plyer==1.4.3
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyarrow==5.0.0
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyparsing==2.4.7
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
pyrsistent==0.18.0
PySide2==5.15.2
PySimpleGUI==4.45.0
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2021.1
pywebio==1.3.3
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
pyzmq==22.2.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
QtPy==1.9.0
requests==2.26.0
rethinkdb==2.4.7
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.7.1
shiboken2==5.15.2
six==1.15.0
sklearn==0.0
smmap==4.0.0
streamlit==0.86.0
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.5.1
zipp==3.5.0
